add: USkinnedMeshComponent, USkeletalMeshComponent(아직 빌드는 안댐)

+++ NipsEngine/Source/Engine/Render/Resource/VertexTypes.h
@@ -32,7 +32,8 @@ struct FSkeletalMeshVertex
     FVector4 Tangent;
 
     // Bone influence (핵심)
-    int32 BoneIndices[4]; // -1로 influence 없음 표시, 256개 이상의 Bone을 가진 Mesh 대응
+    // TODO: invalid bone 표현이나 256개 이상의 bone 표현을 위해 int32로 변경을 고려
+    uint8 BoneIndices[4];
     float BoneWeights[4];
 };
